requirements.txt
Django==4.2
pycryptodome==3.20.0
sympy==1.12
